+++ src/app/workspace/external_data-variable_name_to_interpolation_type_mapping.R
@@ -1,0 +1,5 @@
+mrfei_variable_name_to_interpolate_by_sum_boolean_mapping <- c(TRUE)
+names(mrfei_variable_name_to_interpolate_by_sum_boolean_mapping) <- c("mRFEI")
+
+walkability_variable_name_to_interpolate_by_sum_boolean_mapping <- c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE)
+names(walkability_variable_name_to_interpolate_by_sum_boolean_mapping) <- c("COUNTHU10", "TOTPOP10", "HH", "WORKERS", "AC_TOT", "AC_WATER", "AC_LAND", "AC_UNPR", "D2A_EPHHM", "D2B_E8MIXA", "D3b", "D4a", "D2A_Ranked", "D2B_Ranked", "D4A_Ranked", "D3B_Ranked", "NatWalkInd", "Shape_Length", "Shape_Area")