+++ src/app/workspace/tests/tester.R
@@ -1,5 +1,4 @@
 library(testthat)
 test_file(path='tests/test_acmt_network.R')
 test_file(path='tests/test_external_data.R')
-
 test_file(path='tests/test_example.R')  # not really a test for ACMT, but for ensuring examples run successfully and can demonstrate something